// src/types/appsSettings.ts
export type AppsSettings = {
    id: number;
    customerAppMinimumBuildNumberAndroid: number;
    customerAppMinimumBuildNumberIos: number;
    customerAppUrlAndroid: string;
    customerAppUrlIos: string;
    providerAppMinimumBuildNumberAndroid: number;
    providerAppMinimumBuildNumberIos: number;
    providerAppUrlAndroid: string;
    providerAppUrlIos: string;
    deliveryAppMinimumBuildNumberAndroid: number;
    deliveryAppMinimumBuildNumberIos: number;
    deliveryAppUrlAndroid: string;
    deliveryAppUrlIos: string;
}